refactor: add all sound commands back, hide most under owner only
