+++ src/commands/soundCommands/erb.js
@@ -1,0 +1,13 @@
+const { GenericVoiceCommand } = require('../../models/')
+
+module.exports = new GenericVoiceCommand({
+  triggers: ['erb', 'epicrapbattles', 'rap'],
+  description: 'Play a random epic rap battle of history',
+
+  reaction: '🎤',
+  dir: 'erb',
+  ext: 'opus',
+  np: true,
+  skipIfPlaying: true,
+  ownerOnly: true
+})
