+++ src/commands/soundCommands/fart.js
@@ -5,7 +5,6 @@ module.exports = new GenericVoiceCommand({
   description: 'ew!',
 
   existingConn: 'I only have one butthole, dude. Please wait until the current sound is done or the ear-poo ghost will visit you in your sleep!',
-  reaction: '💩',
   dir: 'farts',
   ext: 'opus',
   message: 'I recorded this myself. Enjoy. 🌬'
