+++ src/commands/soundCommands/hey.js
@@ -1,12 +1,13 @@
 const { GenericVoiceCommand } = require('../../models/')
 
 module.exports = new GenericVoiceCommand({
-  triggers: ['theone', 'thememe'],
-  description: 'I AM THE ONE',
+  triggers: ['hey', 'HEYYEYAAEYAAAEYAEYAA'],
+  description: 'HEYYEYAAEYAAAEYAEYAA I SAID HEY',
 
   existingConn: 'I can only talk so much my dude, wait until I\'m done with whatever sound is playing before trying',
   reaction: '👍',
-  dir: 'theone',
+  dir: 'hey',
   ext: 'opus',
-  files: 'theone'
+  files: 'hey',
+  ownerOnly: true
 })
