+++ src/commands/soundCommands/mememusic.js
@@ -1,11 +1,10 @@
 const { GenericVoiceCommand } = require('../../models/')
 
 module.exports = new GenericVoiceCommand({
-  triggers: ['erb', 'epicrapbattles', 'rap'],
-  description: 'Play a random epic rap battle of history',
+  triggers: ['mememusic', 'memesound', 'shitsound'],
+  description: 'Meme music? More like bad music',
 
-  reaction: '🎤',
-  dir: 'erb',
+  dir: 'shitsound',
   ext: 'opus',
   np: true,
   skipIfPlaying: true
