+++ src/commands/soundCommands/mlgmusic.js
@@ -4,7 +4,6 @@ module.exports = new GenericVoiceCommand({
   triggers: ['mlg', 'mlgmusic'],
   description: 'You want some sweet mlg remixes kids?',
 
-  reaction: '😃',
   dir: 'mlg',
   ext: 'opus',
   files: 25,
