+++ src/commands/soundCommands/oof.js
@@ -8,5 +8,6 @@ module.exports = new GenericVoiceCommand({
   existingConn: 'I only have voice, dude. Please wait until the current sound is done, you assbutt',
   dir: 'oof',
   ext: 'opus',
-  files: 'oof'
+  files: 'oof',
+  ownerOnly: true
 })
