+++ src/commands/soundCommands/playclip.js
@@ -4,5 +4,6 @@ module.exports = new GenericSoundboardCommand({
   triggers: ['playclip'],
   description: 'Plays a custom sound clip',
   usage: '{command} <clipname>',
-  existingConn: 'Develop the technology to speak two different things simultaneously and then we\'ll talk.'
+  existingConn: 'Develop the technology to speak two different things simultaneously and then we\'ll talk.',
+  ownerOnly: true
 })
